PocketMine.php: refactor timezone handling into its own class This removes lots of bloat from PocketMine.php, and this is also important for reusability across threads.
